Modified the default GitHub MIT License to include correct license holders
Added Robert Nystrom as a copyright holder because he wrote the book Crafting Interpreters from which this code is derived.  2015 is the date listed on the GitHub repo’s licese by Robert Nystrom as of this writing.
Added Cortis Clark because I have modified the code to the Dart language as a part of working on Book 6 for the Computer Science Book Club @ Meta