Modified project.pbxproj to open in Xcode 12.5
I created this in Xcode 14 on a computer I no longer have access to.  Xcode 12.5.1, the latest version I can run on my current computer refuses to open it.  I dug around a bit in project.pbxproj and changed the objectVersion (I set it to 54) and other stuff until it would open without complaint.  Xcode actually changed some of the numbers (objectVersion changed from my setting of 54 to 55) and compatibility version changed from 12.5 to 13.0.  It now opens an runs without warnings in Xcode 12.5.1